+++ front/app/containers/Admin/settings/customize/index.tsx
@@ -43,7 +43,8 @@ import { isNilOrError } from 'utils/helperUtils';
 // i18n
 import { InjectedIntlProps } from 'react-intl';
 import { FormattedMessage, injectIntl } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from './messages';
+import sharedSettingsMessages from '../messages';
 
 // services
 import { localeStream } from 'services/locale';
@@ -582,14 +583,14 @@ class SettingsCustomizeTab extends PureComponent<
         <form onSubmit={this.save}>
           <Section key={'branding'}>
             <SectionTitle>
-              <FormattedMessage {...messages.titleHomepageStyle} />
+              <FormattedMessage {...messages.brandingTitle} />
             </SectionTitle>
             <SectionDescription>
-              <FormattedMessage {...messages.subtitleHomepageStyle} />
+              <FormattedMessage {...messages.brandingDescription} />
             </SectionDescription>
 
             <SubSectionTitle>
-              <FormattedMessage {...messages.titlePlatformBranding} />
+              <FormattedMessage {...messages.colorsTitle} />
             </SubSectionTitle>
 
             {['color_main', 'color_secondary', 'color_text'].map(
@@ -642,9 +643,9 @@ class SettingsCustomizeTab extends PureComponent<
             )}
 
             <SectionField key={'logo'}>
-              <Label htmlFor="tenant-logo-dropzone">
+              <SubSectionTitle>
                 <FormattedMessage {...messages.logo} />
-              </Label>
+              </SubSectionTitle>
               <ImagesDropzone
                 id="tenant-logo-dropzone"
                 acceptedFileTypes="image/jpg, image/jpeg, image/png, image/gif"
@@ -660,21 +661,24 @@ class SettingsCustomizeTab extends PureComponent<
           </Section>
 
           <Section key={'header'}>
-            <SubSectionTitle>
+            <SectionTitle>
               <FormattedMessage {...messages.header} />
-            </SubSectionTitle>
+            </SectionTitle>
+            <SectionDescription>
+              <FormattedMessage {...messages.headerDescription} />
+            </SectionDescription>
             <Outlet
               id="app.containers.Admin.settings.customize.headerSectionStart"
               latestAppConfigSettings={latestAppConfigSettings}
               handleOnChange={this.handleSettingOnChange}
             />
             <SectionField key={'header_bg'}>
-              <Label htmlFor="landingpage-header-dropzone">
+              <SubSectionTitle>
                 <FormattedMessage {...messages.header_bg} />
                 <IconTooltip
                   content={<FormattedMessage {...messages.header_bgTooltip} />}
                 />
-              </Label>
+              </SubSectionTitle>
               <StyledImagesDropzone
                 id="landingpage-header-dropzone"
                 acceptedFileTypes="image/jpg, image/jpeg, image/png, image/gif"
@@ -693,7 +697,10 @@ class SettingsCustomizeTab extends PureComponent<
               />
             </SectionField>
 
-            <SectionField>
+            <SectionField key={'banner_text'}>
+              <SubSectionTitle>
+                <FormattedMessage {...messages.bannerTextTitle} />
+              </SubSectionTitle>
               <InputMultilocWithLocaleSwitcher
                 type="text"
                 valueMultiloc={latestAppConfigCoreSettings?.['header_title']}
@@ -738,7 +745,10 @@ class SettingsCustomizeTab extends PureComponent<
                 )}
               />
             </SectionField>
-            <SectionField>
+            <SectionField key="avatars">
+              <SubSectionTitle>
+                <FormattedMessage {...messages.avatarsTitle} />
+              </SubSectionTitle>
               <Setting>
                 <ToggleLabel>
                   <StyledToggle
@@ -763,12 +773,12 @@ class SettingsCustomizeTab extends PureComponent<
           </Section>
 
           <Section key={'project_header'}>
-            <SubSectionTitle>
+            <SectionTitle>
               <FormattedMessage {...messages.projects_header} />
-              <IconTooltip
-                content={formatMessage(messages.projects_header_tooltip)}
-              />
-            </SubSectionTitle>
+            </SectionTitle>
+            <SectionDescription>
+              <FormattedMessage {...messages.projectsHeaderDescription} />
+            </SectionDescription>
             <SectionField>
               <InputMultilocWithLocaleSwitcher
                 type="text"
@@ -783,9 +793,12 @@ class SettingsCustomizeTab extends PureComponent<
           </Section>
 
           <Section>
-            <SubSectionTitle>
+            <SectionTitle>
               <FormattedMessage {...messages.homePageCustomizableSection} />
-            </SubSectionTitle>
+            </SectionTitle>
+            <SectionDescription>
+              <FormattedMessage {...messages.homePageCustomizableDescription} />
+            </SectionDescription>
 
             <WideSectionField>
               <QuillMultilocWithLocaleSwitcher
@@ -845,10 +858,10 @@ class SettingsCustomizeTab extends PureComponent<
             loading={this.state.loading}
             status={getSubmitState({ errors, saved, diff: attributesDiff })}
             messages={{
-              buttonSave: messages.save,
-              buttonSuccess: messages.saveSuccess,
-              messageError: messages.saveErrorMessage,
-              messageSuccess: messages.saveSuccessMessage,
+              buttonSave: sharedSettingsMessages.save,
+              buttonSuccess: sharedSettingsMessages.saveSuccess,
+              messageError: sharedSettingsMessages.saveErrorMessage,
+              messageSuccess: sharedSettingsMessages.saveSuccessMessage,
             }}
           />
         </form>