+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## UNRELEASED
+## Unreleased
+
+/
+
+## 2021-12-08
+
+### Changed
+
+- Improved the structure and copy of the Admin > Settings > Customize page.
+
+### Fixed
+
+- Insights scan category button no longer appears when the insights nlp feature flag is disabled
 
 ## 2021-11-30